requirements.txt
openai>=1.0.0
requests>=2.28.0
ffmpeg-python>=0.2.0
google-api-python-client>=2.60.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
python-dotenv>=0.20.0 